progress: AssetManager; SkillsUSA is awesome!!!

+++ Magma/src/Magma/UI/UIRenderer.h
@@ -22,7 +22,7 @@ public:
 	static void BeginFrame();
 	static void EndFrame();
 
-	static void Pop(uint32_t count = 1);
+	static void Pop(uint64_t count = 1);
 
 	static UIState DrawWindow(Window& window);
 	static UIState DrawButton(Button& button);
